Made the page less laggy, and a bit cleaner.
Added a logarithmic multiple-of-n distance algorithm to decide which
scroll watchers should be active at any time.
Also moved the render pass into a one-time parse-script-and-render when
the script is loaded, instead of re-parsing the whole script on every
redraw.
Also improved the debug flag to allow hiding all debug information.
Also made the text somewhat nicer to read, using multiple layers of
blurred shadows to get a good looking white-with-black-border.