Added support for multiple image versions.
This properly enables mature & h content.
All such content is strictly hidden until the user enables this in the settings menu; the images are handled separately & will refuse to load.  This is to prevent an unsuspecting user from being directed to a NSFW page.

Also fixed a bug that prevented from "original" soundtracks from loading.
Also fixed a bug that was causing animations not to play when audio failed to load.